+++ Bio/Medline/__init__.py
@@ -210,3 +210,9 @@ def read(handle):
     """
     records = parse(handle)
     return next(records)
+
+
+if __name__ == "__main__":
+    from Bio._utils import run_doctest
+
+    run_doctest()